change cleaning to liquid control
