+++ libgui/CMakeLists.txt
@@ -25,8 +25,8 @@ ucm_add_files(
   "movement-window.cpp"
   "manual-movement-window.cpp"
   "status-window.cpp"
-  "cleaning-window.cpp"
-  "cleaning-control-window.cpp"
+  "liquid-status-window.cpp"
+  "liquid-control-window.cpp"
   "plc-trigger-window.cpp"
   "speed-profile-window.cpp"
   "system-info-window.cpp"
