+++ libgui/liquid-status-window.cpp
@@ -1,6 +1,6 @@
 #include "gui.hpp"
 
-#include "cleaning-window.hpp"
+#include "liquid-status-window.hpp"
 
 #include "util.hpp"
 
@@ -8,14 +8,14 @@ NAMESPACE_BEGIN
 
 namespace gui {
 
-CleaningWindow::CleaningWindow(float                   width,
-                               float                   height,
-                               const ImGuiWindowFlags& flags)
-    : Window{"Cleaning", width, height, flags} {}
+LiquidStatusWindow::LiquidStatusWindow(float                   width,
+                                       float                   height,
+                                       const ImGuiWindowFlags& flags)
+    : Window{"Liquid Status", width, height, flags} {}
 
-CleaningWindow::~CleaningWindow() {}
+LiquidStatusWindow::~LiquidStatusWindow() {}
 
-void CleaningWindow::show([[maybe_unused]] Manager* manager) {
+void LiquidStatusWindow::show([[maybe_unused]] Manager* manager) {
   unsigned int status_id = 0;
 
   const ImVec2 size = util::size::h_wide(32.0f);
